removing factory
