+++ src/main/java/com/google/cloud/tools/appengine/cloudsdk/process/DefaultProcessExitListener.java
@@ -16,11 +16,13 @@
 
 package com.google.cloud.tools.appengine.cloudsdk.process;
 
+import com.google.cloud.tools.appengine.cloudsdk.CloudSdk;
+
 /**
  * Default process exit listener that simply captures the exit code and makes it available with a
  * getter. Before the process exit code is captured, it's initialized to -1.
  *
- * <p>When used with {@link com.google.cloud.tools.appengine.cloudsdk.internal.sdk.CloudSdk.Builder}, this
+ * <p>When used with {@link CloudSdk.Builder}, this
  * allows checking of the success or failure of an action execution.
  */
 public class DefaultProcessExitListener implements ProcessExitListener {
