+++ src/test/java/com/google/cloud/tools/appengine/cloudsdk/CloudSdkAppEngineInstancesTest.java
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.appengine.cloudsdk.internal;
+package com.google.cloud.tools.appengine.cloudsdk;
 
 import com.google.cloud.tools.appengine.api.instances.DefaultInstancesSelectionConfiguration;
 import com.google.cloud.tools.appengine.cloudsdk.internal.process.ProcessRunnerException;
-import com.google.cloud.tools.appengine.cloudsdk.internal.sdk.CloudSdk;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
