+++ src/test/java/com/google/cloud/tools/appengine/cloudsdk/CloudSdkAppEngineVersionsTest.java
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.appengine.cloudsdk.internal;
+package com.google.cloud.tools.appengine.cloudsdk;
 
 import com.google.cloud.tools.appengine.api.versions.DefaultVersionsListConfiguration;
 import com.google.cloud.tools.appengine.api.versions.DefaultVersionsSelectionConfiguration;
+import com.google.cloud.tools.appengine.cloudsdk.CloudSdkAppEngineVersions;
 import com.google.cloud.tools.appengine.cloudsdk.internal.process.ProcessRunnerException;
-import com.google.cloud.tools.appengine.cloudsdk.internal.sdk.CloudSdk;
+import com.google.cloud.tools.appengine.cloudsdk.CloudSdk;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
